fix(mouth): an internal page was answering 200 on the public domain
While proving PR #6313 served, I measured something neither it nor #6277 had
closed, and reported it rather than filing the row as done: three INTERNAL
`(workspace)` pages were `"use client"` and each held a hardcoded staff table — a
photo map keyed by email prefix, a tax-consultant dropdown of email and display
name. A client module's constants ship to the browser, so the names of the two
people the owner excluded from public pages sat in three route chunks. The PAGES
are behind auth; the CHUNKS are not. A Next static asset is served from the public
CDN path with no session, so an anonymous curl of
`/_next/static/chunks/app/(workspace)/lkpm/page-*.js` returned 200 and both names.
"Internal page" does not cover a public asset, and calling those chunks an
"authenticated internal surface" in #6313's pack was my sentence and it was wrong.

The tables move to a server module and reach the pages as required props. The photo
map is DERIVED from the roster — measured against the hand-written one it replaces:
14 entries identical, 0 disagreeing, 2 email aliases the roster cannot express kept
explicitly, and 4 people who have a roster portrait but were missing from the map,
who now render it. The tax list is MOVED and deliberately NOT derived: its values
are constrained by a backend CHECK and the roster DISAGREES with two of them —
`faysha.tax@` against `faisha.tax@`, `tax@balizero.com` against `veronika.tax@` —
so deriving would have silently changed what the form submits. Which record is
wrong is the owner's call, and a test now pins both spellings so a later tidy-up
into a derivation fails instead of picking a side.

Moving data to server props put it in the RSC payload inside the HTML, and for one
route that was WORSE than the chunk: `/lkpm` answered 200 on the public domain with
48,529 bytes. The cause was older and larger than the tables — `src/proxy.ts` keeps
an INTERNAL_ROUTES list that redirects workspace pages off the public domain, and
`/lkpm` was the only one missing from it while `/team-management` and `/clients`
already answered 301. One array element fixes that, and two tests pin it. After it,
every anonymous public-host variant of /lkpm — plain, RSC header, prefetch header,
`?_rsc=` — answers 301, and the public response is 30 bytes with no names.

The guard from #6313 is now wired into `npm run build` (scripts.build only; no
dependency, package-lock byte-identical), tightened from "public routes, workspace
skipped by design" to every chunk with an empty allowlist, and extended to scan
`public/` filenames — where the two portraits still sit, now listed in
ACCEPTED_PUBLIC_FILES with the reason, so that exposure is a decision a reader can
see and a third such portrait fails the build.

Measured: 530 chunks, 0 carrying either name, down from 3. Build exit 0 with the
guard inside it; a reachable regression exits 1.
