+++ apps/mouth/package.json
@@ -4,7 +4,7 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "LLMS_GENERATE_ARTICLES_ONLY=1 tsx scripts/generate-llms-full.ts && next build --webpack && node scripts/assert-public-login-bundle.mjs",
+    "build": "LLMS_GENERATE_ARTICLES_ONLY=1 tsx scripts/generate-llms-full.ts && next build --webpack && node scripts/assert-public-login-bundle.mjs && node scripts/assert-roster-not-in-public-chunks.mjs",
     "start": "next start",
     "typecheck": "tsc --noEmit",
     "lint": "eslint .",
