+++ apps/mouth/src/proxy.ts
@@ -19,6 +19,7 @@ const INTERNAL_ROUTES = [
   "/second-home",
   "/settings",
   "/team-management", // workspace team management (not /team which is public)
+  "/lkpm", // workspace LKPM batch — answered 200 on the public domain until 2026-09-12
   "/whatsapp",
   "/admin",
   "/agents",
