+++ evidence/2026-09/agent-mini-pro2-mouth-shweb-w2-c4-workspace-2026-e4b976dd/journal.jsonl
@@ -1,0 +1,4 @@
+{"seat":"kimi-code/k3","role":"review","ok":true,"ts":"2026-09-12T13:48:00Z","round":"POST","verdict":"GO-WITH-CONDITIONS","record":"reviews/refuter-kimi-k3.txt","scope":"the changed hunks of the C4 diff, delivered inline"}
+{"seat":"gemini-3.1-pro","role":"review","ok":true,"ts":"2026-09-12T13:54:00Z","round":"POST","verdict":"NO-GO","record":"reviews/refuter-gemini-3.1-pro.txt","scope":"the same, independently"}
+{"seat":"tp1-qwen3.8-max","role":"review","ok":true,"ts":"2026-09-12T14:26:00Z","round":"POST","verdict":"GO","record":"reviews/council-tp1-qwen3.8-max.txt","scope":"the C4 core after option A: the server module, the guard, the three wrappers, the routing element"}
+{"seat":"codex-gpt-5.6-sol","role":"review","ok":false,"ts":"2026-09-12T14:26:00Z","round":"POST","verdict":null,"record":"reviews/codex-dead-on-mini.txt","scope":"DEAD ON MINI since 2026-09-11: acct1 answers 'workspace out of credits' with exit 0, acct2 times out in the tool router. Recorded, not retried."}
